requirements.txt
streamlit
pandas
autogen-agentchat~=0.2
openai
requests